Add DiceRollSource trait
This trait is used by the code in parser as a soruce of randomness instead of
a raw rand::Rng. This is so that at test time an alternative implementation
can be swapped in. An implementation that uses the gen_range function on
rand::Rng is provided as the default.